move recovery word viewmodel.
